no me acuerdo ni que hicimos

+++ src/HTTPRequest.h
@@ -217,7 +217,7 @@ struct http_request {
     char headers[MAX_HEADERS_LENGTH];
     in_port_t             dest_port;
     char header_host[MAX_FQDN];
-    uint16_t header_content_length;
+    int header_content_length;
 };
 
 struct http_parser {
